+++ src/plone/restapi/types/adapters.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 """JsonSchema providers."""
+from plone.autoform.interfaces import WIDGETS_KEY
+
 from plone.app.textfield.interfaces import IRichText
 from zope.component import adapter
 from zope.component import getMultiAdapter
@@ -30,7 +32,7 @@ from zope.schema.interfaces import ITuple
 from zope.schema.interfaces import IVocabularyFactory
 
 from plone.restapi.types.interfaces import IJsonSchemaProvider
-from plone.restapi.types.utils import get_fieldsets
+from plone.restapi.types.utils import get_fieldsets, get_tagged_values
 from plone.restapi.types.utils import get_jsonschema_properties
 
 
@@ -249,12 +251,17 @@ class ChoiceJsonSchemaProvider(DefaultJsonSchemaProvider):
         choices = []
         enum = []
         enum_names = []
-        if self.field.vocabularyName:
+        if getattr(self.field, 'vocabularyName', None):
             vocabulary = getUtility(
                 IVocabularyFactory,
                 name=self.field.vocabularyName)(self.context)
-        else:
+        elif getattr(self.field, 'vocabulary', None):
             vocabulary = self.field.vocabulary
+        else:
+            tagged = get_tagged_values([self.field.interface], WIDGETS_KEY)
+            vocab_name = tagged[self.field.getName()].get('vocabulary', None)
+            vocab_fac = getUtility(IVocabularyFactory, name=vocab_name)
+            vocabulary = vocab_fac(self.context)
 
         if IContextSourceBinder.providedBy(vocabulary):
             vocabulary = vocabulary(self.context)
@@ -370,3 +377,9 @@ class DatetimeJsonSchemaProvider(DateJsonSchemaProvider):
 
     def get_widget(self):
         return 'datetime'
+
+
+@adapter(ITuple, Interface, Interface)
+@implementer(IJsonSchemaProvider)
+class SubjectsFieldJsonSchemaProvider(ChoiceJsonSchemaProvider):
+    pass